Create Game, Board and Piece classes/subclasses.
The Piece class and subclasses represent each of the different types of pieces in chess as well as the rules pertaining to their movement.

The Board class represents the 8x8 game board that chess is played on.

The Game class represents a game of chess and contains methods for manipulating the game pieces on the game board, detecting the current state of the game, tracking whose turn it is to move, and detecting checkmate/stalemates.