Update the move method of the Chess class.
This method checks whether or not a player can make a move (determined by the current game state, whether the arguments passed are out of range, and if move adhere's to that piece's rules). For example, if game is already finished or player is trying to move one of the opponent's pieces then obviously move is invalid.