Add server.py and fix castling bug
Fix a bug in the is_in_mate function of game.py where castling when a king is checkmated is considered a valid move and thus does not cause the game state to update from 'UNFINISHED' to 'WHITE WON' or 'BLACK WON'. Now castling is no longer considered a valid way to escape checkmate.